Fix support of container URL
